Reorganized test cases into individual files. Added a custom test suite to run all the unit test cases. Added tests for DataSource, SaveHDF.
Added framework for requirements and auto-generated test document output.
