+++ riglib/experiment/experiment.py
@@ -631,16 +631,18 @@ class Experiment(traits.HasTraits, threading.Thread):
         system at end of task. The HDF file is re-opened and any extra task 
         data kept in RAM is written
         '''
-        traits = self.class_editable_traits()
+        if hasattr(self, "h5file"):
+            traits = self.class_editable_traits()
 
-        if hasattr(tables, 'open_file'): # function name depends on version
-            h5file = tables.open_file(self.h5file.name, mode='a')   
-        else:
-            h5file = tables.openFile(self.h5file.name, mode='a')
-        for trait in traits:
-            if (trait not in self.object_trait_names): # don't save traits which are complicated python objects to the HDF file    # and (trait not in ['bmi', 'decoder', 'ref_trajectories']):
-                h5file.root.task.attrs[trait] = getattr(self, trait)
-        h5file.close()
+            if hasattr(tables, 'open_file'): # function name depends on version
+                h5file = tables.open_file(self.h5file.name, mode='a')   
+            else:
+                h5file = tables.openFile(self.h5file.name, mode='a')
+
+            for trait in traits:
+                if (trait not in self.object_trait_names): # don't save traits which are complicated python objects to the HDF file    # and (trait not in ['bmi', 'decoder', 'ref_trajectories']):
+                    h5file.root.task.attrs[trait] = getattr(self, trait)
+            h5file.close()
 
     def end_task(self):
         '''
